Option to skip koi install
We are unable to add additional command line options to the rials new command that get passed into the template. This approach allows us to ask if we would like koi installed as we have been using this template for our non koi based projects as we want to leverage the work to align our projects to Katalyst best practices.

Perhaps if we go down this path we should change the repo name?